Fix issues with editor being mispositioned between saves.
This commit fixes an issue where if you started a new post, then saved, then reloaded thepage, the editor would be pushed down an extra 28px compared to an empty editor.

The specific fix involved unsetting `overflow: hidden;` on `.layout__content`. Magically, this fixed everything.

It's a doozy of a bug, because it appears, aside from content entered in TinyMCE, that there was zero difference in the rendered markup between an empty editor, and an editor with a saved post. So why was there a 28px discrepancy? After spending an hour doing web-inspector surgery, I would wager the error stems from one of two categories:

1. It's an obtuse issue with complicated CSS that involves perspectives, animations, translates and absolute and fixed positioned elements, perhaps with a browser bug thrown in for good measure. Position fixed and iframes (TinyMCE lives in an iframe) are usually recipes for "success" in this vein. Interestingly, despite the TinyMCE container having a height of 800px in both an empty editor, and an editor with a saved post, unsetting `overflow:hidden;` in both caused a vertical scrollbar to appear in one of these, but not the other.

2. It's an issue with paddings and margins on a stray element or classname that is unique to an editor editing a saved post, vs. an empty edotor, which I haven't catched.

In case you read this and some of the ingredients ring a bell, please chime in. In case it's the 2nd category of bugs, that means this commit doesn't actually "fix" the issue, but rather covers it up. CC: @mtias